restapi/requirements.txt
flask
cassandra-driver

